refactor out TextEditor.tsx
